requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth